+++ src/components/producers_competition.jl
@@ -11,6 +11,7 @@ module ProducersCompetition_
 include("blueprint_modules.jl")
 include("blueprint_modules_identifiers.jl")
 import .EN: Foodweb, _Foodweb
+import .EN: Topologies as G
 
 #-------------------------------------------------------------------------------------------
 mutable struct Raw <: Blueprint
@@ -30,7 +31,21 @@ function F.late_check(raw, bp::Raw)
 end
 
 F.expand!(raw, bp::Raw) = expand!(raw, bp.alpha)
-expand!(raw, alpha) = raw._scratch[:producers_competition] = alpha
+function expand!(raw, alpha)
+    # Only add trophic edges where there are non-*missing* values (not *non-zero*).
+    # This way, user can modify zero values later if they were non-missing.
+    sources, targets, _ = findnz(alpha)
+    mask = spzeros(Bool, size(alpha))
+    for (src, tgt) in zip(sources, targets)
+        mask[src, tgt] = true
+    end
+    raw._scratch[:producers_competition] = alpha
+    raw._scratch[:producers_competition_mask] = mask
+    g = raw._topology
+    ety = :producers_competition
+    G.add_edge_type!(g, ety)
+    G.add_edges_within_node_type!(g, :species, ety, mask)
+end
 
 #-------------------------------------------------------------------------------------------
 mutable struct Flat <: Blueprint
@@ -138,17 +153,27 @@ function (::_ProducersCompetition)(alpha = nothing; kwargs...)
 
 end
 
+@propspace producers.competition
+
 @expose_data edges begin
-    property(producers.competition)
+    property(producers.competition.matrix)
     depends(ProducersCompetition)
     @species_index
     ref(raw -> raw._scratch[:producers_competition])
-    get(ProducersCompetitionRates{Float64}, sparse, "producers link")
-    template(raw -> @ref raw.producers.matrix)
+    get(ProducersCompetitionMatrix{Float64}, sparse, "producers competition rates matrix")
+    template(raw -> @ref raw.producers.competition.mask)
     write!((raw, rhs::Real, i, j) -> ProducersCompetition_.check(rhs, (i, j)))
+end
+
+@expose_data edges begin
+    property(producers.competition.mask)
+    depends(ProducersCompetition)
+    @species_index
+    ref(raw -> raw._scratch[:producers_competition_mask])
+    get(ProducersCompetitionMask{Bool}, sparse, "producers competition edges mask")
 end
 
 function F.shortline(io::IO, model::Model, ::_ProducersCompetition)
     print(io, "Producers competition: ")
-    showrange(io, model.producers._competition)
+    showrange(io, model.producers.competition._matrix)
 end